Embed spinner definitions
The original code worked by reading a json file that it assumed was located
in a directory relative to the parent directory.  This approach is brittle
and was problematic for methods like PyInstaller that try to bundle all the
modules used by an application.  To compound the problem, the json files
were located in a directory named "cli-spinners", which is not a valid
Python module name (because of the hyphen).  This made it difficult to find
a way to use Python import to grab the necessary files explicitly when
using PyInstaller.  After trying and failing with various other approaches,
I finally gave up and grabbed the contents from the cli-spinners repository
on GitHub: https://github.com/sindresorhus/cli-spinners

cli-spinners is Copyright (C) Sindre Sorhus and MIT licensed.